chart rendering and localstorage logic cleanup
